Minor changes. Prepare for VisualStudio compatibility. Remove some less relevant info from the periodic stat printout. Example batch files for solo and pool mining , disable Windows Error Reporting before running the miner to be automatically restart if it crashes. Re-enable sending WPS statistic to the pool in the  windows miner. (balance /payout still disabled)

+++ bignum.h
@@ -117,7 +117,7 @@ public:
 
     unsigned int getuint() const
     {
-        return BN_get_word(this);
+        return (unsigned int)BN_get_word(this);
     }
 
     int getint() const
